DjangoBackend/requirements.txt
# Backend & API
Django==5.1.3
djangorestframework==3.15.2
gunicorn==20.1.0
uvicorn==0.15.0

# Autentificare
dj-rest-auth==7.0.1
django-allauth==65.9.0
PyJWT>=2.5.0

# Bază de date & Cache
psycopg2-binary==2.9.10
redis==6.0.0

# Task Queue (Celery)
celery==5.5.3
django-celery-beat==2.8.1
django-redis

# Cloud & Stocare (AWS, Azure, Firebase)
boto3==1.39.4
django-storages==1.14.6
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.32.0
firebase-admin==6.5.0

# Google Services & AI
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googlemaps==4.10.0

# Utilitare
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.13.4
blinker==1.7.0
cachetools==5.5.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cryptography==3.4.8
django-cors-headers==4.7.0
exceptiongroup==1.2.2

geographiclib==2.0
geopy==2.4.1
httplib2==0.20.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
oauthlib==3.2.0
pillow==11.2.1
protobuf>=3.19.5,<5.0.0
pydantic==2.9.2
pytz==2025.2
Pygments==2.16.1
PyMuPDF==1.24.13
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.2
starlette==0.46.1
tqdm==4.66.6
urllib3==2.0.7
Werkzeug==3.0.1
pandas

# --- AI & LangChain (Versiuni compatibile) ---
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.8

# --- PDF & Căutare Web ---
pypdf==4.2.0
duckduckgo-search==5.3.0
tavily-python==0.5.0

# --- Data & Excel (Pandas) ---
openpyxl==3.1.2
xlsxwriter==3.2.0

cloudinary
